Create SQL Project Planning
REQUEST
1.	There is only 1 table, TASK, each row has start and end, when the end  dates are consecutive it means is the same project. I have to count the length of each project.
a.	Order table by end dates, bring previous row date in a new column with a CTE, then subtract New column to End Date, if is consecutive result will be 0, if not result will be more
b.	Then I have to group those in different projects numbers, with a running total using a windows function, since the ones that are consecutive have 0, it wont add up, until a new project appears, this will make the separation for projects.